+++ src/containers/ConsultationRequest.tsx
@@ -1,6 +1,7 @@
 import { useEffect, useState } from 'react';
 import Section from '../components/Section';
 import { type TRequestedConsultationsProps } from '../lib/types';
+import { RequestedConsultaionItem } from '../components/RequestedConsultaionItem';
 
 type TConsultationRequestProps = {
   onApprove: (consultation: TRequestedConsultationsProps) => void;
@@ -58,61 +59,19 @@ export default function ConsultationRequest({
     );
   };
 
-  // 빠른 상담 요청 여부에 따른 테두리 변경
-  const getBorderColorClass = (quick: boolean) => {
-    return quick ? 'quick-border' : 'border-gray-200';
-  };
-
   return (
     <>
       <Section title='들어온 상담 요청'>
         <div className='w-full h-fit p-4'>
           {consultationData.length > 0 ? (
-            consultationData.map(
-              ({
-                id,
-                name,
-                title,
-                hopeDay,
-                hopeTime,
-                requestDay,
-                approvalStatus,
-                quick,
-              }) => (
-                <article
-                  key={id}
-                  className='w-full flex flex-col justify-end mb-4 items-start'
-                >
-                  <small className='ml-2'>{requestDay}</small>
-                  {quick}
-                  <div
-                    className={`bg-white rounded-lg border ${getBorderColorClass(quick)}  p-4 shadow-lg w-full`}
-                  >
-                    <div className='flex justify-between items-center'>
-                      <div className='flex flex-col w-[70%]'>
-                        <span className='w-[95%] text-[1rem] font-bold truncate	'>
-                          {name} 손님
-                        </span>
-                        <span className='w-[95%] font-bold truncate	'>
-                          {title}
-                        </span>
-                        <span className='text-[0.8rem] font-bold'>
-                          요청일: {hopeDay} {hopeTime}
-                        </span>
-                      </div>
-
-                      <button
-                        className='w-[6rem] h-[2rem] text-white text-[1rem] rounded-lg bg-hanaindigo hover:bg-hanagold'
-                        onClick={() => toggleApprovalStatus(id)}
-                      >
-                        {!approvalStatus && '승인대기'}
-                      </button>
-                    </div>
-                  </div>
-                </article>
-              )
-            )
-          ) : (
+          consultationData.map((consultation) => (
+            <RequestedConsultaionItem
+              key={consultation.id}
+              {...consultation}
+              onApprove={toggleApprovalStatus}
+            />
+          ))
+        ) : (
             <div className='text-center text-hanaindigo p-4 h-full text-xl border-x border-b border-gray-200'>
               모든 상담 요청이 승인되었습니다.
             </div>
